refactor: :recycle: change structure folder
