+++ src/components/CloseButton/index.tsx
