+++ src/components/FeedbackContentStep/types.ts
@@ -1,0 +1,8 @@
+import { FeedbackType } from "components/WidgetForm/types"
+
+export interface FeedbackContentStepProps {
+  feedbackType: FeedbackType
+  onFeedbackRestartRequested: () => void
+  onFeedbackSent: () => void
+}
+
