+++ src/components/FeedbackSuccessStep/types.ts
@@ -1,0 +1,3 @@
+export interface FeedbackSuccessStepProps {
+  onFeedbackRestartRequested: () => void
+}
