+++ src/components/FeedbackTypeStep/types.ts
@@ -1,0 +1,5 @@
+import { FeedbackType } from "components/WidgetForm/types";
+
+export interface FeedbackTypeStepProps {
+  onFeedbackTypeChanged: (type: FeedbackType) => void
+}
