+++ src/components/ScreenshotButton/types.ts
@@ -1,0 +1,4 @@
+export interface ScreenshotButtonProps {
+  screenshot: string | null
+  onScreenshotTook: (screenshot: string | null) => void
+}
