+++ src/components/WidgetForm/index.tsx
@@ -1,8 +1,8 @@
 import { useState } from 'react'
-import { FeedbackTypeStep } from './Steps/FeedbackTypeStep'
-import { FeedbackContentStep } from './Steps/FeedbackContentStep'
+import { FeedbackTypeStep } from '../FeedbackTypeStep'
+import { FeedbackContentStep } from '../FeedbackContentStep'
+import { FeedbackSuccessStep } from '../FeedbackSuccessStep/FeedbackSuccessStep'
 import { FeedbackType } from './types'
-import { FeedbackSuccessStep } from './Steps/FeedbackSuccessStep'
 
 const WidgetForm: React.FC = () => {
   const [feedbackType, setFeedbackType] = useState<FeedbackType | null>(null)
