+++ src/components/WidgetForm/types.ts
@@ -1,22 +1,3 @@
-import feedbackTypes from './feedbackTypes'
+import feedbackTypes from 'data/feedbackTypes'
 
 export type FeedbackType = keyof typeof feedbackTypes
-
-export interface FeedbackTypeStepProps {
-  onFeedbackTypeChanged: (type: FeedbackType) => void
-}
-
-export interface FeedbackContentStepProps {
-  feedbackType: FeedbackType
-  onFeedbackRestartRequested: () => void
-  onFeedbackSent: () => void
-}
-
-export interface ScreenshotButtonProps {
-  screenshot: string | null
-  onScreenshotTook: (screenshot: string | null) => void
-}
-
-export interface FeedbackSuccessStepProps {
-  onFeedbackRestartRequested: () => void
-}
